+++ src/main/java/org/interledger/codecs/oer/ipr/InterledgerPaymentRequestOerCodec.java
@@ -1,11 +1,11 @@
 package org.interledger.codecs.oer.ipr;
 
+import org.interledger.Condition;
 import org.interledger.InterledgerAddress;
 import org.interledger.codecs.Codec;
 import org.interledger.codecs.CodecContext;
 import org.interledger.codecs.InterledgerPaymentRequestCodec;
 import org.interledger.codecs.oer.OerUint8Codec.OerUint8;
-import org.interledger.ilp.Condition;
 import org.interledger.ilp.InterledgerPayment;
 import org.interledger.ipr.InterledgerPaymentRequest;
