+++ src/main/java/org/interledger/ledger/LedgerAdaptor.java
@@ -1,8 +1,8 @@
 package org.interledger.ledger;
 
+import org.interledger.Fulfillment;
 import org.interledger.InterledgerAddress;
 import org.interledger.InterledgerException;
-import org.interledger.ilp.Fulfillment;
 import org.interledger.ledger.events.LedgerConnectEvent;
 import org.interledger.ledger.events.LedgerEvent;
 import org.interledger.ledger.events.LedgerEventHandler;
